:recycle: refactor(patients): Merge two hooks.
  - usePatientsSuggestions
  - makePatientsSuggestions
